passing lints and tests. Client interface incomplete and examples mising
